backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
yfinance>=0.2.55
curl_cffi>=0.7.0
pandas==2.2.3
anthropic>=0.40.0,<1.0.0
httpx>=0.27.0
python-dotenv==1.0.1
